Alex L - Setting up component with todo service
