+++ src/App.js
@@ -1,70 +1,45 @@
-import React, { useState } from 'react';
-import logo from './logo.svg';
+import React, { useEffect, useState } from 'react';
+import todoApi from './api/todoApi';
 import './App.css';
-import { getAllByPlaceholderText } from '@testing-library/react';
-
-const Todo = ({ todo }) => <div className="todo">{todo.name}</div>;
-
-function TodoForm({ addTodo }) {
-  const [value, setValue] = useState("");
-
-  const handleSubmit = e => {
-    e.preventDefault();
-    if (!value) return;
-    addTodo(value);
-    setValue("");
-  };
-
-  return (
-    <form onSubmit={handleSubmit}>
-      <input
-        type="text"
-        className="input"
-        value={value}
-        onChange={e => setValue(e.target.value)}
-      />
-    </form>
-  );
-}
+import TodoForm from './components/TodoForm';
 
 function App() {
-  const [todos, setTodos] = useState([
-    {
-      id: 1,
-      name: "Learn about React",
-      isCompleted: false,
-      description: "  I want to learn about react and stuff",
-      date: new Date().toLocaleString()
-    },
-    {
-      id: 2,
-      name: "Meet friend for lunch",
-      isCompleted: false,
-      description: "A FRIEND AND I ARE GOING TO GET PASTA",
-      date: new Date().toLocaleString()
-    },
-    {
-      id: 3,
-      name: "Build a really cool todo app",
-      isCompleted: false,
-      description: "the app that youre looking at must be prettier",
-      date: new Date().toLocaleString()
-    }
-  ]);
+  // useState is a hook that helps you manage the state (i.e. the values in the component).
+  // useState returns two things 1. the data, 2. a function to update the data
+  // in this case the state is the `todos` and the way  to update the state is using `setTodos`
+  const [todos, setTodos] = useState([])
 
-
+  // useEffect is called when the component is being `mounted`
+  useEffect(() => {
+    // load the todos from the server
+    const data = todoApi.getAll();
+    // update the state of the component -> will cause the componet to redraw themselves
+    setTodos(data);
+  }, []);
 
   const addTodo = name => {
-    const newTodos = [...todos, { name }];
-    setTodos(newTodos);
+    const newTodo = {
+      name, // same as writing `name: name`
+      description: "Missing desc",
+      isCompleted: false
+    }
+    // Call todoApi.add to update the todos
+    todoApi.add(newTodo);
+
+    // Get the latest todos and update the state;
+    // The reason we use the ... operator is to copy the data over
+    // so that react can "notice" there was a change
+    setTodos([...todoApi.getAll()]);
   };
 
+  // TODO: Alex L - try to redo the logic using todoApi
   const completeTodo = index => {
     const newTodos = [...todos];
     newTodos[index].isCompleted = true;
     setTodos(newTodos);
   };
 
+  // TODO: Alex L - try to redo the logic using todoApi
   const removeTodo = index => {
     const newTodos = [...todos];
     newTodos.splice(index, 1);
@@ -72,14 +47,24 @@ function App() {
   };
 
   function Todo({ todo, index, completeTodo, removeTodo }) {
+
+    function KeyValue({ label, children }) {
+      return (
+        <div>
+          <span style={{ fontWeight: "bold" }}>{label}: </span>{children}
+        </div>
+      );
+    }
+
     return (
       <div
         className="todo"
         style={{ textDecoration: todo.isCompleted ? "line-through" : "" }}>
-        {todo.id}
-        {todo.name}
-        {todo.description}
-        {todo.date}
+        <div className="todo-content">
+          <KeyValue label="ID">{todo.id}</KeyValue>
+          <KeyValue label="Name">{todo.name}</KeyValue>
+          <KeyValue label="Description">{todo.description}</KeyValue>
+        </div>
         <div>
           <button onClick={() => completeTodo(index)}>Complete</button>
           <button onClick={() => removeTodo(index)}>x</button>
